finished basic layout
