+++ index.html
@@ -8,7 +8,7 @@
     @import url("https://cdnjs.cloudflare.com/ajax/libs/normalize/4.2.0/normalize.min.css");
   </style>
   <body>
-  <div id="react-root"></div>
+    <div id="react-root"></div>
   </body>
   <script src="/dist/bundle.js"></script>
 </html>
