+++ src/Header.jsx
@@ -1,0 +1,14 @@
+import React, {Component} from 'react';
+
+class Header extends Component {
+  render() {
+    return (
+    <div>
+      <div className="header">
+        <img src="../img/simdrone_logo.png" id= "logo" />
+      </div>
+    </div>
+    )
+  }
+}
+export default Header;
