+++ src/Video.jsx
@@ -1,0 +1,14 @@
+import React, {Component} from 'react';
+
+class Video extends Component {
+  render() {
+    return (
+    <div>
+      <div className="video">
+        video goes here!
+      </div>
+    </div>
+    )
+  }
+}
+export default Video;
