+++ src/app.html
@@ -12,7 +12,6 @@
 
   <body>
     <div class="container">
-
       <div class="main">
 
         <div class="header">
@@ -56,62 +55,9 @@
         <div class= "request-control"> REQUEST CONTROL </div>
         <div class= "queue-container"> current queue </div>
         <div class= "chat-container"> chatting app </div>
-
       </div>
 
     </div>
-
-
-
-
-
-
-   <!--  <section class="main-container">
-      <nav>
-        <img src="../img/simdrone_logo.png" id= "logo">
-      </nav>
-
-      <div class= "video-control-panel">
-        <div class= "video-container">
-          video
-        </div>
-
-        <div class= "control-panel">
-          <h2 class="control-panel-line"><span> Control Panel </span></h2>
-
-          <div class="control-panel-container">
-            <div class="button-container rotated-buttons">
-              <button class="button"><span class="label">Move Forward</span></button>
-              <button class="button"><span class="label">Move Right</span></button>
-              <button class="button"><span class="label">Move Left</span></button>
-              <button class="button"><span class="label">Move Backward</span></button>
-            </div>
-
-            <div class="center-button-container">
-              <div class="center-button"> Take Off </div>
-              <div class="center-button"> Land </div>
-            </div>
-
-            <div class="button-container rotated-buttons">
-              <button class="button"><span class="label">Rise</span></button>
-              <button class="button"><span class="label">Turn Right</span></button>
-              <button class="button"><span class="label">Turn Left</span></button>
-              <button class="button"><span class="label">Lower</span></button>
-            </div>
-          </div>
-        </div>
-      </div>
-
-      <aside class= "side-bar">
-        <div class= "request-control"> REQUEST CONTROL </div>
-        <div class= "queue-container"> current queue </div>
-        <div class= "chat-container"> chatting app </div>
-      </aside>
-
-    </section>
-
- -->
-
   </body>
 
 </html>
