+++ styles/app.scss
@@ -1,3 +1,4 @@
+
 .container{
   display:flex;
   position:fixed;
@@ -30,6 +31,7 @@
 .controls{
   height: 300px;
   background: white;
+  bottom: 0;
 }
 
 .sidebar{
@@ -138,9 +140,23 @@
 
 .request-control {
   background: grey;
-  width: 100px;
+  width: 250px;
   height: 50px;
-  margin: 20px
+  margin: 20px;
+  text-align: center;
+}
+
+.queue-container {
+  background: grey;
+  width: 300px;
+  height: 200px;
+  margin: 10px 0;
+}
+
+.chat-container {
+  background: grey;
+  width: 300px;
+  height: 400px
 }
