+++ styles/home.scss
@@ -1,91 +1,230 @@
-    body {
-      background: white;
-      color: black;
-      font-family: 'Open Sans', sans-serif;
-      font-size: 14px;
-      margin: 70px 0;
-    }
 
-    .chatbar {
-      background: tomato;
-      border-top: 1px solid black;
-      bottom: 0;
-      display: flex;
-      left: 0;
-      padding: 5px;
-      position: fixed;
-      right: 0;
-    }
+.container{
+  display:flex;
+  position:fixed;
+  top:0;
+  bottom:0;
+  left:0;
+  right:0;
+}
 
-    .chatbar-message,
-    .chatbar-username {
-      border: 1px solid black;
-      display: block;
-      height: 40px;
-      line-height: 40px;
-      margin: 5px;
-      padding: 0 9px;
-    }
+.main{
+  flex-grow: 1;
+  height: 100%;
+  display:flex;
+  flex-direction:column;
+}
 
-    .chatbar-message {
-      flex: 8;
-    }
+.sidebar{
+  height:100%;
+  max-width: 35%;
+  width:40%;
+  background:deepskyblue;
+}
 
-    .chatbar-username {
-      flex: 2;
-    }
+.chat-container {
+  position: fixed;
+  bottom: 0;
+  height: 50%;
+}
 
-    .message {
-      display: flex;
-      padding: 5px;
-    }
+.request-control {
+  background-color: tomato;
+  text-align: center;
+  padding: 20px;
+}
 
-    .message.system {
-      font-style: italic;
-      padding: 5px 20px;
-    }
+#logo {
+  width: 25vh
+}
 
-    .message-content,
-    .message-username {
-      overflow: hidden;
-      margin: 5px;
-      padding: 0 10px;
-    }
+.video{
+  flex-grow: 1;
+  background-color:mediumslateblue;
+}
 
-    .message-content {
-      flex: 8;
-    }
+.control-panel-line {
+  width: 100%;
+  text-align: left;
+  border-bottom: 1px solid #000;
+  line-height: 0.1em;
+  margin: 10px 0 20px
+}
 
-    .message-username {
-      flex: 2;
-      font-weight: 700;
-    }
+.control-panel-line span {
+  background: #fff;
+  padding: 0 10px;
+}
 
-    .navbar {
-      background: tomato;
-      border-bottom: 1px solid black;
-      height: 60px;
-      left: 0;
-      padding: 0 10px;
-      position: fixed;
-      right: 0;
-      top: 0;
-    }
+.control-panel-container {
+  display: flex;
+  justify-content: space-around;
+}
 
-    .navbar-brand {
-      color: inherit;
-      display: block;
-      float: left;
-      font-weight: 700;
-      font-size: 28px;
-      height: 60px;
-      line-height: 60px;
-      padding: 0 10px;
-      text-decoration: none;
-    }
-    .counter{
-      float:right;
-      margin-top:1.3em;
-    }
+.button-container {
+  display: flex;
+  flex-wrap: wrap;
+  width: 240px;
+}
+
+.center-button-container {
+  text-align: center
+}
+
+.center-button {
+  color: white;
+  background: grey;
+  padding: 27px;
+  margin-top: 46px
+}
+
+
+.rotated-buttons {
+  transform: rotate(45deg);
+  padding: 45px;
+}
+
+.rotated-buttons .button {
+  background: #666;
+  border: 0;
+  box-sizing: border-box;
+  color: #fff;
+  display: block;
+  font-size: 13px;
+  height: 90px;
+  margin: 5px;
+  padding: 20px;
+  position: relative;
+  text-transform: uppercase;
+  width: 90px;
+}
+
+.rotated-buttons .button .label {
+  display: block;
+  position: relative;
+  transform: rotate(-45deg);
+}
+
+.rotated-buttons .button:after {
+  background: #aaa;
+  content: '';
+  height: 10px;
+  left: 10px;
+  position: absolute;
+  top: 10px;
+  width: 40px;
+}
+
+.rotated-buttons .button:before {
+  background: #aaa;
+  content: '';
+  height: 40px;
+  left: 10px;
+  position: absolute;
+  top: 10px;
+  width: 10px;
+}
+.rotated-buttons .button:nth-child(2):after,
+.rotated-buttons .button:nth-child(2):before,
+.rotated-buttons .button:nth-child(4):after,
+.rotated-buttons .button:nth-child(4):before {
+  left: auto;
+  right: 10px;
+}
+
+.rotated-buttons .button:nth-child(3):after,
+.rotated-buttons .button:nth-child(3):before,
+.rotated-buttons .button:nth-child(4):after,
+.rotated-buttons .button:nth-child(4):before {
+  bottom: 10px;
+  top: auto;
+}
+
+.controls{
+  position: fixed;
+  bottom: 0;
+}
+
+
+
+
+
+.chatbar {
+  background: tomato;
+  border-top: 1px solid black;
+  bottom: 0;
+  display: flex;
+  padding: 5px;
+  position: fixed;
+  width: 100%;
+}
+
+.chatbar-message,
+.chatbar-username {
+  display: block;
+  margin: 5px;
+  padding: 0 9px;
+}
+
+.chatbar-message {
+  flex: 8;
+}
+
+.chatbar-username {
+  flex: 2;
+}
+
+// .message {
+//   display: flex;
+//   padding: 5px;
+// }
+
+// .message.system {
+//   font-style: italic;
+//   padding: 5px 20px;
+// }
+
+// .message-content,
+// .message-username {
+//   overflow: hidden;
+//   margin: 5px;
+//   padding: 0 10px;
+// }
+
+// .message-content {
+//   flex: 8;
+// }
+
+// .message-username {
+//   flex: 2;
+//   font-weight: 700;
+// }
+
+// .navbar {
+//   background: tomato;
+//   border-bottom: 1px solid black;
+//   height: 60px;
+//   left: 0;
+//   padding: 0 10px;
+//   position: fixed;
+//   right: 0;
+//   top: 0;
+// }
+
+// .navbar-brand {
+//   color: inherit;
+//   display: block;
+//   float: left;
+//   font-weight: 700;
+//   font-size: 28px;
+//   height: 60px;
+//   line-height: 60px;
+//   padding: 0 10px;
+//   text-decoration: none;
+// }
+// .counter{
+//   float:right;
+//   margin-top:1.3em;
+// }
